[DotNetUnitTest Package] Enhanced the DotNetUnitTest Package with a "MbUnit advanced configuration" mode that enables the use of a set of optional package properties to run the tests using the EXEC task to run MbUnit.Cons.exe, which supports several command line arguments that the original implementation does not (e.g. excluding certain tests from the test run using filters).
NOTE: Using this advanced mode MAY make the build more susceptible to "hanging", since CCNET and/or nant seem to wait indefinitely for the EXECed process to stop (so if it doesn't exit, the build may hang.) However, it does allow for a great deal more control over how the build's unit test behavior, which can a requirement in some situations (e.g. when configuring multiple build types that each need to run different subsets of the test suite).