[PARTIAL] [NCover Package] [NCover v3 Compatibility] Preliminary commit of the reverse-engineered source code for the NCover v3.x  NAnt Tasks.
NOTE: This new code isn't included in the project (i.e. isn't compiled) yet, since additional code changes to existing CI Factory source have yet to be integrated with this new code.

NOTE: The NAnt Task DLLs that come with NCover are compiled against the mainline version of NAnt, and are incompatible with the CI Factory version of NAnt due to a difference in the return type of the ExternalProgramBase.BaseDirectory property. Therefore, equivalent NAnt Tasks have to be compiled using the CI Factory version of NAnt in order for them to be usable in CI Factory NAnt scripts.